requirements.txt
web3>=6.0.0
python-dotenv>=0.19.0
requests>=2.26.0
aiohttp>=3.8.0
colorama>=0.4.4